% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{est.riskfxn}
\alias{est.riskfxn}
\title{a function to compute the risk function}
\usage{
est.riskfxn(t, delta, ci = F)
}
\arguments{
\item{t}{a vector of times}

\item{delta}{a vector of event indicators}

\item{ci}{indicator of whether the function should output 95\% CI on risk function (default is F)}
}
\value{
dataframe with risk function and standard error at each observed event time
}
\description{
a function to compute the risk function
}
\examples{
est.riskfxn(times, indicators)



}
