% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.risk}
\alias{get.risk}
\title{a function to output risk at specific timepoints}
\usage{
get.risk(time, risk, taus, se = NULL, output_se = F)
}
\arguments{
\item{time}{a vector of times}

\item{risk}{a vector of risk estimates}

\item{taus}{a vector of timepoints of interest.}

\item{output_se}{an indicator of whether to output standard error (T/F)}
}
\description{
a function to output risk at specific timepoints
}
\examples{
get.risk(obsrisk$time, obsrisk$risk, taus = 4*365.25)


}
