% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{np.cor}
\alias{np.cor}
\title{a function to implement nonparametric correction}
\usage{
np.cor(times, obsfxn, afxn, bfxn)
}
\arguments{
\item{times}{a vector of ordered unique observed event times}

\item{obsfxn}{a vector of risk estimates of each time}

\item{afxn}{a vector of values of a(t)}

\item{bfxn}{a vector of values of b(t)}
}
\value{
dataframe with times and corrected risks
}
\description{
a function to implement nonparametric correction
}
\examples{
np.cor(obsrisk$time, obsrisk$risk, a, b)

}
